Improved performance for group merging
Instead of iterating over all nodes in a group and changing their
leader, instead I point the leader to the new “super” group and added a
getGroup() method that follows the chain of leaders until the top one
is found.

Basically I found that maintaining and updating the child nodes within
a group became very expensive (computationally and memory-wise) at
large node numbers and this fix solves the problem at the expense of
worsening the performance of inSameGroup(). However, I think this is
the right trade-off for how I suspect this data structure may be used.